Extracted Dialog from ImportsScreen
